+++ internal/orders/scheme.go
@@ -28,7 +28,7 @@ type OrderScheme struct {
 	OwnerID    int32     `json:"-"`
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
-	Accrual    float64   `json:"accrual,omitempty"`
+	Accrual    float32   `json:"accrual,omitempty"`
 	UploadedAt time.Time `json:"uploaded_at"`
 }
 
@@ -57,5 +57,5 @@ func (orderList *OrderListScheme) ScanRow(row di.Row) error {
 type AccrualScheme struct {
 	OrderNumber string  `json:"order"`
 	Status      string  `json:"status"`
-	Amount      float64 `json:"accrual,omitempty"`
+	Amount      float32 `json:"accrual,omitempty"`
 }
